src/main/java/ephyl/model/Course.java
package ephyl.model;

import ephyl.dto.StudentDto;

import java.util.List;
import java.util.Objects;

public class Course {
    private int id;
    private String name;
    private List<StudentDto> students;

    private int teacher_id;

    public Course(String name, int teacher_id) {
        this.name = name;
        this.teacher_id = teacher_id;}

    public int getTeacher_id() {
        return teacher_id;
    }

    public void setTeacher_id(int teacher_id) {
        this.teacher_id = teacher_id;
    }

    public List<StudentDto> getStudents() {
        return students;
    }

    public void setStudents(List<StudentDto> students) {
        this.students = students;
    }

    public Course() {
    }

    public Course(int id, String name) {
        this.id = id;
        this.name = name;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Course course = (Course) o;
        return id == course.id && teacher_id == course.teacher_id && Objects.equals(name, course.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name, teacher_id);
    }
}
